# src/async_api_client/openapi/openapi-3.0.3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Airflow API (Stable)",
...
    "version": "2.10.5",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "contact": {
      "name": "Apache Software Foundation",
      "url": "https://airflow.apache.org",
      "email": "dev@airflow.apache.org"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Apache Airflow Stable API."
    }
  ],
  "paths": {
...
    "/config": {
      "get": {
        "summary": "Get current configuration",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.config_endpoint",
        "operationId": "get_config",
        "tags": [
          "Config"
        ],
        "parameters": [
          {
            "name": "section",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "If given, only return config of this section."
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Config"
                },
                "example": {
                  "sections": [
                    {
                      "name": "core",
                      "options": [
                        {
                          "key": "dags_folder",
                          "value": "/home/user/my-dags-folder"
                        }
                      ]
                    },
                    {
                      "name": "smtp",
                      "options": [
                        {
                          "key": "smtp_host",
                          "value": "localhost"
                        },
                        {
                          "key": "smtp_mail_from",
                          "value": "airflow@example.com"
                        }
                      ]
                    }
                  ]
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "[core]\ndags_folder = /home/user/my-dags-folder\n[smtp]\nsmtp_host = localhost\nsmtp_mail_from =  airflow@example.com\n"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigOption": {
        "type": "object",
        "description": "The option of configuration.",
        "properties": {
          "key": {
            "type": "string",
            "readOnly": true
          },
          "value": {
            "type": "string",
            "readOnly": true
          }
        }
      },
      "ConfigSection": {
        "type": "object",
        "description": "The section of configuration.",
        "properties": {
          "name": {
            "type": "string",
            "readOnly": true
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigOption"
            }
          }
        }
      },
      "Config": {
        "type": "object",
        "description": "The configuration.",
        "properties": {
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigSection"
            }
          }
        }
      },
...
      "Error": {
        "description": "[RFC7807](https://tools.ietf.org/html/rfc7807) compliant response.\n",
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI reference [RFC3986] that identifies the problem type. This specification\nencourages that, when dereferenced, it provide human-readable documentation for\nthe problem type.\n"
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type."
          },
          "status": {
            "type": "number",
            "description": "The HTTP status code generated by the API server for this occurrence of the problem."
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence of the problem."
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem. It may or may\nnot yield further information if dereferenced.\n"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
    },
...
    "responses": {
...
      "Unauthenticated": {
        "description": "Request not authenticated due to missing, invalid, authentication info.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PermissionDenied": {
        "description": "Client does not have sufficient permission.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "A specified resource is not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "Basic": {
        "type": "http",
        "scheme": "basic"
      },
      "GoogleOpenId": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://accounts.google.com/.well-known/openid-configuration"
      },
      "Kerberos": {
        "type": "http",
        "scheme": "negotiate"
      }
    }
  },
  "security": [],
...
}